+++ trunk/src/com/dsatab/data/modifier/LeModificator.java
@@ -14,9 +14,9 @@ import com.dsatab.data.enums.AttributeType;
 
 public class LeModificator extends AbstractModificator {
 
-	public static final double LEVEL_1 = 0.5;
-	public static final double LEVEL_2 = 0.33;
-	public static final double LEVEL_3 = 0.25;
+	public static final float LEVEL_1 = 0.5f;
+	public static final float LEVEL_2 = 0.33f;
+	public static final float LEVEL_3 = 0.25f;
 
 	public LeModificator(Hero hero) {
 		super(hero);
@@ -25,7 +25,7 @@ public class LeModificator extends AbstractModificator {
 	@Override
 	public String getModificatorName() {
 
-		double ratio = hero.getLeRatio();
+		float ratio = hero.getLeRatio();
 		if (ratio < LEVEL_3) {
 			return "Lebensenergie < 1/4";
 		} else if (ratio < LEVEL_2) {
@@ -41,7 +41,7 @@ public class LeModificator extends AbstractModificator {
 	@Override
 	public String getModificatorInfo() {
 		String info = null;
-		double ratio = hero.getLeRatio();
+		float ratio = hero.getLeRatio();
 		if (ratio < LEVEL_3) {
 			info = "Eigenschaften, Kampf -3; Talente,Zauber -9; GS -3";
 		} else if (ratio < LEVEL_2) {
@@ -57,7 +57,7 @@ public class LeModificator extends AbstractModificator {
 	public Modifier getModifier(Probe probe) {
 
 		int modifier = 0;
-		double ratio = hero.getLeRatio();
+		float ratio = hero.getLeRatio();
 
 		if (probe instanceof Attribute) {
 			Attribute attribute = (Attribute) probe;
@@ -88,7 +88,7 @@ public class LeModificator extends AbstractModificator {
 	@Override
 	public Modifier getModifier(AttributeType type) {
 		int modifier = 0;
-		double ratio = hero.getLeRatio();
+		float ratio = hero.getLeRatio();
 
 		if (ratio < LEVEL_3) {
 			if (AttributeType.isEigenschaft(type) || AttributeType.isFight(type)) {
